pyproject.toml
[project]
name = "mzcld-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dockerflow>=2024.4.2",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "opentelemetry-distro>=0.53b1",
    "opentelemetry-exporter-otlp>=1.32.1",
    "opentelemetry-instrumentation-fastapi>=0.53b1",
    "opentelemetry-instrumentation-logging>=0.53b1",
    "pydantic-settings>=2.9.1",
    "python-json-logger>=3.3.0",
]
